requirements.txt
Pillow == 3.4.2
django-crispy-form == 1.6.1
django-ipware == 1.1.6